+++ x2engine/js/X2GridView/x2gridview.js
@@ -850,7 +850,6 @@ $.widget("x2.gvSettings", {
         });
     },
     _compareGridviewSettings:function(self) {
-        //console.log ('_compareGridviewSettings');
         self.options.DEBUG && console.log ('this = ');
         self.options.DEBUG && console.log (this);