+++ x2engine/protected/config/X2Config.php
@@ -5,8 +5,8 @@ $host = 'localhost';
 $user = 'root';
 $pass = '';
 $dbname = 'x2engine';
-$version = '5.0.3';
-$buildDate = '1422995918';
+$version = '5.0.4';
+$buildDate = '1423192854';
 $updaterVersion = '5.1.1';
 $language='en';
 ?>