+++ x2engine/protected/modules/actions/views/actions/view.php
@@ -80,10 +80,14 @@ if(isset($associationModel) && $model->associationType=='contacts') {
 }
 
 $this->widget('X2WidgetList', array ('model' => $model));
+
+$completeUrl = $this->createUrl ('complete', array (
+	'id' => $model->id
+)); 
 ?>
 
 <div class="form" id="action-form">
-	<form id="complete-action" name="complete-action" action="complete/<?php echo $model->id; ?>" method="POST">
+	<form id="complete-action" name="complete-action" action="<?php echo $completeUrl ?>">
 		<b><?php echo Yii::t('actions','Completion Notes'); ?></b>
 		<textarea name="note" rows="4" ></textarea>
 	<div class="row buttons">