+++ x2engine/protected/modules/docs/data/install.sql
@@ -21,7 +21,7 @@ CREATE TABLE `x2_docs` (
 INSERT INTO `x2_modules`
 (`name`, title, visible, menuPosition, searchable, editable, adminOnly, custom, toggleable)
 VALUES
-('docs', 'Docs', 1, 10, 0, 0, 0, 0, 0);
+('docs', 'Docs', 1, 9, 0, 0, 0, 0, 0);
 /*&*/
 INSERT INTO `x2_fields`
 (`modelName`, `fieldName`, `attributeLabel`, `modified`, `custom`, `type`, `required`, `readOnly`, `linkType`, `searchable`, `relevance`, `isVirtual`,`keyType`)